Changed fire code to float from double

+++ src/Fire.hpp
@@ -36,13 +36,13 @@ class Fire {
         Fire(URBInputData*, URBGeneralData*, Output*);
         
         struct FireProperties {
-	  double  w, h, d, r, T, tau, K, H0, U_c, L_c;
+	  float  w, h, d, r, T, tau, K, H0, U_c, L_c;
         };
         
         struct FireState {
-            double burn_time;
-            double burn_flag;
-            double front_flag;
+            float burn_time;
+            float burn_flag;
+            float front_flag;
         };
         
         struct FireCell {
@@ -52,44 +52,44 @@ class Fire {
         };
         
         double time=0;
-        double r_max = 0;
-        double dt=0;
+        float r_max = 0;
+        float dt=0;
         
         std::vector< FireCell > fire_cells;
         
-        std::vector<double> w_base;
+        std::vector<float> w_base;
 
-        std::vector<double> front_map;
+        std::vector<float> front_map;
 
-        std::vector<double> del_plus;
+        std::vector<float> del_plus;
 
-        std::vector<double> del_min;
+        std::vector<float> del_min;
 
-        std::vector<double> xNorm;
+        std::vector<float> xNorm;
 
-        std::vector<double> yNorm;
+        std::vector<float> yNorm;
 
-        std::vector<double> Force;
+        std::vector<float> Force;
   
-        std::vector<double> Pot_u;
-        std::vector<double> Pot_v;
-        std::vector<double> Pot_w;
+        std::vector<float> Pot_u;
+        std::vector<float> Pot_v;
+        std::vector<float> Pot_w;
 
   // Potential field
         int pot_z, pot_r, pot_G, pot_rStar, pot_zStar;
         float drStar, dzStar;
-        std::vector<double> u_r;
-        std::vector<double> u_z;
-        std::vector<double> G;
-        std::vector<double> Gprime;
-        std::vector<double> rStar;
-        std::vector<double> zStar;
+        std::vector<float> u_r;
+        std::vector<float> u_z;
+        std::vector<float> G;
+        std::vector<float> Gprime;
+        std::vector<float> rStar;
+        std::vector<float> zStar;
       
         void run(Solver*, URBGeneralData*);
         void move(Solver*, URBGeneralData*);   
 	void potential(URBGeneralData*);   
         void save(Output*);
-        double computeTimeStep();
+        float computeTimeStep();
         
     private:
         
@@ -103,11 +103,11 @@ class Fire {
         float L, W, H, baseHeight, courant;
         int i_start, i_end, j_start, j_end, k_end,k_start;
 
-        double rothermel(FuelProperties*, double, double, double);
+        float rothermel(FuelProperties*, float, float, float);
                      
-        FireProperties balbi(FuelProperties*, double, double, double, double, double, double);
+        FireProperties balbi(FuelProperties*, float, float, float, float, float, float);
         
-        FireProperties runFire(double, double, int);
+        FireProperties runFire(float, float, int);
         
         // output fields
         std::vector<double> burn_flag;
